requirements.txt
bs4
selenium
pytube3